refactor: generator클래스 util패키지로 이동
