+++ src/test/java/kr/codesquad/domain/LadderLineTest.java
@@ -1,5 +1,6 @@
 package kr.codesquad.domain;
 
+import kr.codesquad.util.RandomLineGenerator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
